refactor; test_vis passes
