+++ enclone_visual/src/lib.rs
@@ -37,6 +37,7 @@ pub mod gui_structures;
 pub mod messages;
 pub mod process_messages;
 pub mod style;
+pub mod summary;
 pub mod svg_to_geometry;
 pub mod testsuite;
 pub mod update_restart;
@@ -47,38 +48,6 @@ const SVG_NULL_HEIGHT: u16 = 190;
 const SVG_HEIGHT: u16 = 400;
 
 type MsgFn = fn(Result<(), String>) -> messages::Message;
-
-use iced::{Button, Column, Container, Element, Length, Scrollable, Text};
-use messages::Message;
-pub fn summary(slf: &mut gui_structures::EncloneVisual) -> Element<Message> {
-    let summary_title = Text::new(&format!("Summary")).size(30);
-    let summary = SUMMARY_CONTENTS.lock().unwrap()[0].clone();
-    let nlines = summary.chars().filter(|&n| n == '\n').count();
-    let font_size = (15 * nlines) / 38;
-    let summary_scrollable = Scrollable::new(&mut slf.scroll)
-        .width(Length::Fill)
-        .height(Length::Fill)
-        .scrollbar_width(SCROLLBAR_WIDTH)
-        .scroller_width(12)
-        .style(style::ScrollableStyle)
-        .push(
-            Text::new(&format!("{}", summary))
-                .font(DEJAVU_BOLD)
-                .size(font_size as u16),
-        );
-    let summary_close_button = Button::new(&mut slf.open_state, Text::new("Vanish!"))
-        .on_press(Message::SummaryClose);
-    let content = Column::new()
-        .spacing(SPACING)
-        .padding(20)
-        .push(summary_title)
-        .push(summary_scrollable)
-        .push(summary_close_button);
-    Container::new(content)
-        .width(Length::Fill)
-        .height(Length::Fill)
-        .into()
-}
 
 async fn noop() -> Result<(), String> {
     // Increasing this time to 2000ms will prevent the screen from going dark on initialization
